split qwen36 moe vmm config parsing
